refactor: move package loader to teal.package_loader
