+++ tl.lua
@@ -1,8 +1,5 @@
-local _tl_compat; if (tonumber((_VERSION or ''):match('[%d.]*$')) or 0) < 5.3 then local p, m = pcall(require, 'compat53.module'); if p then _tl_compat = m end end; local assert = _tl_compat and _tl_compat.assert or assert; local ipairs = _tl_compat and _tl_compat.ipairs or ipairs; local load = _tl_compat and _tl_compat.load or load; local package = _tl_compat and _tl_compat.package or package; local string = _tl_compat and _tl_compat.string or string; local table = _tl_compat and _tl_compat.table or table; local type = type
+local _tl_compat; if (tonumber((_VERSION or ''):match('[%d.]*$')) or 0) < 5.3 then local p, m = pcall(require, 'compat53.module'); if p then _tl_compat = m end end; local assert = _tl_compat and _tl_compat.assert or assert; local ipairs = _tl_compat and _tl_compat.ipairs or ipairs; local load = _tl_compat and _tl_compat.load or load; local string = _tl_compat and _tl_compat.string or string; local table = _tl_compat and _tl_compat.table or table; local type = type
 local VERSION = "0.24.6+dev"
-
-local tldebug = require("teal.debug")
-local TL_DEBUG = tldebug.TL_DEBUG
 
 local errors = require("teal.errors")
 
@@ -12,12 +9,7 @@ local lexer = require("teal.lexer")
 
 local types = require("teal.types")
 
-
-
-
-
-
-local a_type = types.a_type
+local package_loader = require("teal.package_loader")
 
 local parser = require("teal.parser")
 
@@ -34,9 +26,6 @@ local type_checker = require("teal.checker.type_checker")
 local type_reporter = require("teal.type_reporter")
 
 
-
-local util = require("teal.util")
-local read_file_skipping_bom = util.read_file_skipping_bom
 
 local environment = require("teal.environment")
 
@@ -137,11 +126,11 @@ local tl = { EnvOptions = {}, TypeCheckOptions = {} }
 
 
 
-
 tl.check = type_checker.check
 tl.search_module = require_file.search_module
 tl.warning_kinds = errors.warning_kinds
 tl.lex = lexer.lex
+tl.loader = package_loader.install_loader
 tl.generate = lua_generator.generate
 tl.get_token_at = lexer.get_token_at
 tl.parse = parser.parse
@@ -204,64 +193,10 @@ tl.gen = function(input, env, opts, parse_lang)
    return code, result
 end
 
-local function tl_package_loader(module_name)
-   local found_filename, fd, tried = tl.search_module(module_name, false)
-   if found_filename then
-      local parse_lang = parser.lang_heuristic(found_filename)
-      local input = read_file_skipping_bom(fd)
-      if not input then
-         return table.concat(tried, "\n\t")
-      end
-      fd:close()
-      local program, errs = tl.parse(input, found_filename, parse_lang)
-      if #errs > 0 then
-         error(found_filename .. ":" .. errs[1].y .. ":" .. errs[1].x .. ": " .. errs[1].msg)
-      end
-
-      local env = tl.package_loader_env
-      if not env then
-         tl.package_loader_env = assert(environment.for_runtime(parse_lang), "Default environment initialization failed")
-         env = tl.package_loader_env
-      end
-      local defaults = env.defaults
-
-      local w = { f = found_filename, x = 1, y = 1 }
-      env.modules[module_name] = a_type(w, "typedecl", { def = a_type(w, "circular_require", {}) })
-
-      local result = tl.check(program, found_filename, defaults, env)
-
-      env.modules[module_name] = result.type
-
-
-
-      local code = assert(tl.generate(program, defaults.gen_target, lua_generator.fast_opts))
-      local chunk, err = load(code, "@" .. found_filename, "t")
-      if chunk then
-         return function(modname, loader_data)
-            if loader_data == nil then
-               loader_data = found_filename
-            end
-            local ret = chunk(modname, loader_data)
-            return ret
-         end, found_filename
-      else
-         error("Internal Compiler Error: Teal generator produced invalid Lua. Please report a bug at https://github.com/teal-language/tl\n\n" .. err)
-      end
-   end
-   return table.concat(tried, "\n\t")
-end
-
-function tl.loader()
-   if package.searchers then
-      table.insert(package.searchers, 2, tl_package_loader)
-   else
-      table.insert(package.loaders, 2, tl_package_loader)
-   end
-end
 
 local function env_for(parse_lang, env_tbl)
    if not env_tbl then
-      return assert(tl.package_loader_env)
+      return assert(package_loader.env)
    end
 
    if not tl.load_envs then
@@ -279,10 +214,10 @@ tl.load = function(input, chunkname, mode, ...)
       return nil, (chunkname or "") .. ":" .. errs[1].y .. ":" .. errs[1].x .. ": " .. errs[1].msg
    end
 
-   if not tl.package_loader_env then
-      tl.package_loader_env = environment.for_runtime(parse_lang)
+   if not package_loader.env then
+      package_loader.env = environment.for_runtime(parse_lang)
    end
-   local defaults = tl.package_loader_env.defaults
+   local defaults = package_loader.env.defaults
 
    local filename = chunkname or ("string \"" .. input:sub(45) .. (#input > 45 and "..." or "") .. "\"")
    local result = tl.check(program, filename, defaults, env_for(parse_lang, ...))
